Update best_model
Resolved a bug in the logic of saving the best model.
Instead of applying 'np.argmax' on the 'best_models' dictionary, the 'max()' function is applied on the extracted output test score (best_models[x][1]) for each model (x) and returns the model name (x) with the highest test score.